Rename tools to annotation.
