+++ vision/tools/blob_annotator/annotation/__init__.py
