+++ vision/tools/blob_annotator/annotation/generate_annotation.py
