+++ vision/tools/blob_annotator/annotation/location_generalizer.py
